moved unsafe things to Toppokki.Unsafe removed unused types
